redesign
dsgo's current public api surface is only served
from root package. This is good for easy adapting
but it makes it hard to understand project semantics.

That's so, packages are re-organized. This is the
init step of redesign. dsgo also provides too
many features within modules and core, this makes
it hard to manage edge/corner cases. All of them added
into the repository without testing with real life
scenario and to make is match with DSPy parity. However,
this project no more following DSPy framework. It's
now pivoted fro the initial architectural implementation.

note: API_KEY management and custom provider support is
removed for now. This is the first minor step to reduce
complexities and future-solidness of the application.

Signed-off-by: assagman <ahmetsercansagman@gmail.com>